refactor: move code from openshift
This is needed to avoid circular package dependencies.
